#mod the problem's results are now int #fix the problem's result string has a fixxed length now

+++ ProjectEuler/src/main/java/problems/LogicBase.java
@@ -5,6 +5,6 @@ public abstract class LogicBase {
   public LogicBase() {
   }
 
-  public abstract double getResult(Number... numbers);
+  public abstract int getResult(Number... numbers);
 
 }
